Solar_Model: update models to take advantage of new functionality
Solar_Model_Areas
-----------------

* [REF] Rebuilt using `solar make-model`.


Solar_Model_Bookmarks
---------------------

* [REF] Rebuilt using `solar make-model`.


Solar_Model_Nodes
-----------------

* [REF] Rebuilt using `solar make-model`.


Solar_Model_Nodes_Record
------------------------

* [REF] Move all significant __getTagsAsString() logic to the
  Solar_Model_Tags_Collection class.

* [REF] Move all significant __setTagsAsString() logic to the
  Solar_Model_Tags_Collection class.

* [DEL] Method _postSave().  The automatic connection of mapping records and
  the new Solar_Model_Tags_Collection logic now takes care of all the previous
  logic.


Solar_Model_Taggings
--------------------

* [REF] Rebuilt using `solar make-model`.


Solar_Model_Tags
----------------

* [REF] Rebuilt using `solar make-model`.


Solar_Model_Tags_Collection
---------------------------

* [ADD] Method setNames() to set all records in the collection to a list of
  tag names.  Automatically adds and removes tags from the collection based
  on the new list of names.

* [ADD] Method getNamesAsString() to return a space-separated list of tag
  names in the collection.

+++ Solar/Model/Nodes/Record.php
@@ -1,21 +1,11 @@
 <?php
 /**
  * 
- * A single record from the "nodes" model.
- * 
- * @category Solar
- * 
- * @package Solar_Model
- * 
- * @author Paul M. Jones <pmjones@solarphp.com>
- * 
- * @license http://opensource.org/licenses/bsd-license.php BSD
- * 
- * @version $Id$
+ * A single Solar_Model_Nodes record.
  * 
  */
-class Solar_Model_Nodes_Record extends Solar_Model_Record
-{
+class Solar_Model_Nodes_Record extends Solar_Model_Record {
+    
     /**
      * 
      * Magic method to get the 'tags_as_string' property.
@@ -25,17 +15,13 @@ class Solar_Model_Nodes_Record extends Solar_Model_Record
      */
     public function __getTagsAsString()
     {
-        // if exactly null, populate for the first time.
+        // populate for the first time
         if (empty($this->_data['tags_as_string'])) {
-            $text = '';
             // $this->tags forces the __get() call to the related object,
             // then only proceeds if there are tags there.
             if ($this->tags) {
-                foreach ($this->tags as $tag) {
-                    $text .= "{$tag->name} ";
-                }
+                $this->_data['tags_as_string'] = $this->tags->getNamesAsString();
             }
-            $this->_data['tags_as_string'] = rtrim($text);
         }
         
         return $this->_data['tags_as_string'];
@@ -45,6 +31,8 @@ class Solar_Model_Nodes_Record extends Solar_Model_Record
      * 
      * Magic method to set the 'tags_as_string' property.
      * 
+     * Maintains the tags collection on-the-fly.
+     * 
      * @param string $val A space-separated list of tags.
      * 
      * @return void
@@ -52,125 +40,11 @@ class Solar_Model_Nodes_Record extends Solar_Model_Record
      */
     public function __setTagsAsString($val)
     {
-        $val = preg_replace('/[ ]{2,}/', ' ', $val);
-        $this->_data['tags_as_string'] = trim($val);
-        $new = explode(' ', $this->_data['tags_as_string']);
-        $new = array_unique($new);
-        
         if (! $this->tags) {
-            $related = $this->_model->getRelated('tags');
-            $this->tags = $related->getModel()->newCollection();
+            $this->tags = $this->newRelated('tags');
         }
-        
-        // build a new tag collection
-        $model = $this->tags->getModel();
-        $coll = $model->newCollection();
-        foreach ($new as $name) {
-            $record = $model->fetchNew();
-            $record->name = $name;
-            $coll[] = $record;
-        }
-        
-        // reset the tags collection
-        $this->tags = $coll;
-    }
-    
-    /**
-     * 
-     * Corrects and saves related records and collections.
-     * 
-     * @return void
-     * 
-     */
-    protected function _postSave()
-    {
-        // -------------------------------------------------------------
-        // 
-        // make sure that each tag actually exists; insert a new tag
-        // record for each new one.
-        // 
-        
-        // hold on to the tags model
-        $tags_model = $this->tags->getModel();
-        
-        // get the list of tags on the record
-        $tag_names = $this->tags->getNames();
-        if ($tag_names) {
-            // make sure each tag actually exists as a record.  tag names
-            // that do not exist will not be in this array.
-            $params = array(
-                'where' => array(
-                    'tags.name IN (?)' => $tag_names,
-                ),
-            );
-            $existing_tags = $tags_model->fetchAll($params);
-        } else {
-            $existing_tags = $tags_model->newCollection(null);
-        }
-        
-        // loop through the tags
-        foreach ($this->tags as $key => $tag) {
-            
-            // does this tag already exist?
-            foreach ($existing_tags as $existing) {
-                if ($tag->name == $existing->name) {
-                    // replace with the existing tag record.
-                    // this lets us "connect" user-entered tag names
-                    // with existing records. capture both the tag from
-                    // the loop and reset the collection value.
-                    $tag = $existing;
-                    $this->tags[$key] = $existing;
-                    break;
-                }
-            }
-            
-            // if status is still 'new' then save it
-            if ($tag->getStatus() == 'new') {
-                $tag->save();
-            }
-            
-            // look for a tagging on this tag
-            $found = false;
-            foreach ($this->taggings as $tagging) {
-                // must match node ID and tag ID
-                $match = $tagging->node_id == $this->id &&
-                         $tagging->tag_id  == $tag->id;
-                if ($match) {
-                    $found = true;
-                    break;
-                }
-            }
-            
-            // did we find a tagging for the tag?
-            if (! $found) {
-                // no, create one ...
-                $tagging = $this->taggings->getModel()->fetchNew();
-                $tagging->node_id = $this->id;
-                $tagging->tag_id  = $tag->id;
-                
-                // ... and add to the collection
-                $this->taggings[] = $tagging;
-            }
-        }
-        
-        // remove taggings that don't exist any more
-        foreach ($this->taggings as $key => $tagging) {
-            $found = false;
-            foreach ($this->tags as $tag) {
-                $match = $tagging->node_id == $this->id &&
-                         $tagging->tag_id  == $tag->id;
-                if ($match) {
-                    $found = true;
-                    break;
-                }
-            }
-            
-            // does the tagging match a tag?
-            if (! $found) {
-                // no, remove from the taggings
-                $tagging->delete();
-            }
-        }
+        $this->tags->setNames($val);
+        $this->_data['tags_as_string'] = $this->tags->getNamesAsString();
     }
     
     /**
@@ -182,6 +56,8 @@ class Solar_Model_Nodes_Record extends Solar_Model_Record
      */
     protected function _postDelete()
     {
-        $this->taggings->delete();
+        if ($this->taggings) {
+            $this->taggings->delete();
+        }
     }
 }
